+++ app/src/main/java/org/ghostsinthelab/apps/guilelessbopomofo/MainActivity.kt
@@ -200,7 +200,7 @@ class MainActivity : AppCompatActivity(), Vibratable {
                         hapticFeedbackPreferenceStrength = progress
                         performVibrate(
                             applicationContext,
-                            hapticFeedbackPreferenceStrength.toLong()
+                            hapticFeedbackPreferenceStrength
                         )
 
                         sharedPreferences.edit()
